Punto 1
Lecturas de Quantopian.
Lecture 1 Introduction to Research A simple tutorial to help you get up to speed in the research environment.
Lecture 2 Introduction to Python Some basic tools for working in the language.
Lecture 3 Introduction to NumPy How to use NumPy for computing on data.
Lecture 4 Introduction to pandas An introduction to using pandas to manage and analyze your data.
Lecture 5 Plotting Data A brief primer.